Update header handling
The header brams used to have one entry for each of 96
slots, with entries re-used for each antenna.
This commit changes the bram addressing so that it
is possible to have a different channel and IP address
for each channel. This should make it easier to support
a 350-antenna spoof mode.