fixed antenna assignment in multisnap
by changing assign_slot to call set_ant_headers as well. To reflect
change in the design, assign_slot now sets the antenna headers as well. Added an option
to disable the ref in SNAPADC so you can drive the SNAP at a lower clock rate if needed.
Removed software register control for nants because it was not really controlling number of antennas.