Catch I2C initialization error
Instantiating the I2C objects requires that the running design
have the approriate I2C registers.
This causes issues when instantiating the correlator class
just to program the boards.
This commit catches errors in the I2C instantiation, and in such an
event will attempt to instantiate them when the Fengine's `initialize`
method is called.