Allow storage of bitstream in redis
Use the hera_upload_fpg_to_redis.py script to load a bitstream to redis
Use the syntax
    fpgfile: redis:<fpgname>
in the config file to supply a bitstream this way.
This should make it easier for multiple machines to control the SNAPs
without requiring shared storage for fpg files or harmonization of path names.

Note that on programming a file to redis it will be temporarily written to
disk on the machine being used for programming, in the /tmp/ directory.
Currently this is hard coded.