Only un-dead programmed F-engines
The CasperFpga.is_connected() check which was being
used to figure out if a dead board has come back to life
will pass if the board is not programmed.
Subsequently, such a board will immediately be re-declared dead
by the status loop.

This commit adds the option for an additional check -- that the
fengine is programmed -- before adding it back into rotation.

The poll loop has been updated to use this check.