Add lane bonding test to ADC initialization
This commit adds a check to the ADC initialization method
to ensure that all ADC lanes are cycle-synchronized.

Occassionally a situation arises where one lane is a clock
cycle out of sync with its interleaving group, leading to
a mangling of the sample ordering.

Probably some more careful process could detect this error and
fix it using appropriate use of the ADC SERDES bitslippers, this
commit does nothing more than allow the situation to be detected,
and flagged as an initilization failure.

This commit will probably lead to more init retries than are really
needed, since nothing in the DSP pipeline requires different
ADC channels to be mutually synchronized. As such, the max number
of retries has been turned up to 5.